Refactor reachability related code in the semantic analyzer
This just moves reachability related code to a new module.
