+++ mypy/semanal_pass1.py
@@ -27,9 +27,8 @@ from mypy.nodes import (
     MODULE_REF, implicit_module_attrs, AssertStmt,
 )
 from mypy.types import Type, UnboundType, UnionType, AnyType, TypeOfAny, NoneTyp, CallableType
-from mypy.semanal import (
-    SemanticAnalyzerPass2, infer_reachability_of_if_statement, assert_will_always_fail,
-)
+from mypy.semanal import SemanticAnalyzerPass2
+from mypy.reachability import infer_reachability_of_if_statement, assert_will_always_fail
 from mypy.semanal_shared import create_indirect_imported_name
 from mypy.options import Options
 from mypy.sametypes import is_same_type
